requirements.txt
argparse>=1.4.0
google-api-python-client
requests
validators
Pillow
